bb-imager-gui: ui: Combine flash finish views
Were very small files, doing similar things.
